Corregido error con los tipo de dato entero

+++ etlstat/extractor/extractor.py
@@ -13,7 +13,7 @@ conversion_map = {
     'STRING': str,
     'NUMBER': np.float32,
     'DECIMAL': np.float32,
-    'INTEGER': np.int32
+    'INTEGER': np.float32
 }
